requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
pytz==2025.1
# Optional: uncomment when you add real providers
# amadeus==8.0.0
# duffel-api==2.8.0
google-generativeai>=0.5.0   # The correct, modern GenAI SDK (real version)
openai>=1.0.0       # Add openai to prevent fallback errors